requirements.txt
python-nmap 
rich 
paramiko 
impacket 
requests 
ldap3 
mysql-connector-python 
psycopg2-binary 
redis
